+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
- - Tests for `ByteRange`, `BytesRepresentation`, `StorePrefix`, `StoreKey`, `ArrayBuilder`, `ArraySubset`, `GroupBuilder`, `Group`, `NodeName`, `NodePath`, `Node`, `AdditionalFields`, `Metadata`, `FillValue`
+ - Tests for `ByteRange`, `BytesRepresentation`, `StorePrefix`, `StoreKey`, `ArrayBuilder`, `ArraySubset`, `GroupBuilder`, `Group`, `NodeName`, `NodePath`, `Node`, `AdditionalFields`, `Metadata`, `FillValue`, `Group`
  - `array_subset::IncompatibleStartEndIndicesError`
  - Add `array::transmute_from_bytes_vec`
  - Re-export public dependencies at the crate root: `bytes`, `bytemuck`, `dyn_clone`, `serde_json`, `ndarray`, `object_store`, and `opendal`
@@ -35,13 +35,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - **Breaking**: `InvalidByteRangeError` now holds a `ByteRange` and bytes length and returns a more informative error message
  - **Breaking**: Remove `StorageError::InvalidJSON` and add `StorageError::InvalidMetadata`
    - `InvalidMetadata` additionally holds a `StoreKey` for more informative error messages
- - **Breaking**: Remove `NodeCreateError::Metadata`
-   - `NodeCreateError::StorageError` with `StorageError::InvalidMetadata` is used instead
 
 ### Removed
  - **Breaking**: Remove `StorePrefixError::new`, deprecated since `v0.7.3`
  - **Breaking**: Remove `ArraySubset::{in_subset,in_subset_unchecked}`, deprecated since `v0.7.2`
  - **Breaking**: Remove `impl From<&StorePrefix> for NodeName`, unused and not useful
+ - **Breaking**: Remove `NodeCreateError::Metadata`
+   - `NodeCreateError::StorageError` with `StorageError::InvalidMetadata` is used instead
+ - **Breaking**: Remove `{ArrayCreateError,GroupCreateError}::MetadataDeserializationError`
+   - `{ArrayCreateError,GroupCreateError}::StorageError` with `StorageError::InvalidMetadata` is used instead
+ - **Breaking**: Remove `GroupCreateError::Metadata` as it was unused
 
 ### Fixed
  - Disallow an empty string for a `StoreKey`